refactor: 버튼 컴포넌트 내용 기능과 스타일로 분리
