+++ src/components/Buttons/DeleteButton.jsx
@@ -1,42 +1,10 @@
-import styled from "styled-components";
-import { DeletedIcon } from "../Icons";
-import { theme } from "../../styles/theme";
-
-const DeleteButtonWrapper = styled.button`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 36px;
-    height: 36px;
-    padding: 6px;  
-    border: 1px solid #CCCCCC;
-    border-radius: 6px;
-    ${(props) => props.$disable ?
-      `background-color: #CCCCCC;` :
-      `background-color: #FFFFFF;
-    
-      &:hover {
-          cursor: pointer;
-          background-color: #F6F6F6;
-      }
-
-      &:focus {
-          border: 1px solid #555555;
-          background-color: #FFFFFF;
-      }`
-    }
-`;
-
-const returnDeletedIcon = (disable) => {
-  const color = (disable ? theme.colors.white : theme.colors.black);
-  return <DeletedIcon color={color} />
-}
+import * as DeleteButtonStyle from "./DeleteButtonStyle";
 
 function DeleteButton({ $disable }) {
     return (
-      <DeleteButtonWrapper $disable={$disable}>
-        {returnDeletedIcon($disable)}
-      </DeleteButtonWrapper>
+      <DeleteButtonStyle.DeleteButtonWrapper $disable={$disable}>
+        {DeleteButtonStyle.returnDeletedIcon($disable)}
+      </DeleteButtonStyle.DeleteButtonWrapper>
     );
   }
